+++ spec/resources/paw/mac_os_x/10_10_spec.rb
@@ -1,18 +1,21 @@
 require_relative '../../../spec_helper'
 
 describe 'resource_paw::mac_os_x::10_10' do
+  let(:name) { 'default' }
   let(:action) { nil }
   let(:runner) do
-    ChefSpec::SoloRunner.new(step_into: 'paw',
-                             platform: 'mac_os_x',
-                             version: '10.10')
+    ChefSpec::SoloRunner.new(
+      step_into: 'paw', platform: 'mac_os_x', version: '10.10'
+    ) do |node|
+      %i(name action).each do |a|
+        node.set['resource_paw_test'][a] = send(a) unless send(a).nil?
+      end
+    end
   end
-  let(:chef_run) do
-    runner.converge("resource_paw_test::#{action}")
-  end
+  let(:chef_run) { runner.converge('resource_paw_test') }
 
   context 'the default action (:install)' do
-    let(:action) { :default }
+    let(:action) { nil }
 
     it 'configures the Mac App Store' do
       expect(chef_run).to include_recipe('mac-app-store')
@@ -24,7 +27,22 @@ describe 'resource_paw::mac_os_x::10_10' do
 
     it 'installs via the App Store' do
       expect(chef_run).to install_mac_app_store_app('Paw (HTTP & REST Client)')
-        .with(bundle_id: 'com.luckymarmot.Paw')
+    end
+  end
+
+  context 'the :upgrade action' do
+    let(:action) { :upgrade }
+
+    it 'configures the Mac App Store' do
+      expect(chef_run).to include_recipe('mac-app-store')
+    end
+
+    it 'upgrades the Paw app' do
+      expect(chef_run).to upgrade_paw('default')
+    end
+
+    it 'upgrades via the App Store' do
+      expect(chef_run).to upgrade_mac_app_store_app('Paw (HTTP & REST Client)')
     end
   end
 end