+++ spec/support/cookbooks/resource_paw_test/recipes/default.rb
@@ -1,3 +1,7 @@
 # Encoding: UTF-8
 
-paw 'default'
+attrs = node['resource_paw_test']
+
+paw attrs['name'] do
+  action attrs['action'] unless attrs['action'].nil?
+end