+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Paw Cookbook CHANGELOG
 
 Unreleased
 ----------
+- Update to use mac-app-store cookbook v2
 
 v0.1.0 (2016-04-27)
 -------------------